requirements.txt
django>=4.0.5
djangorestframework
django-cors-headers
django-extensions
psycopg2-binary
python-dotenv==0.21.0